+++ Changelog.md
@@ -6,6 +6,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+
+09/23/2022
+### Added
+- Redux toolkit for statemanagement
+- Summary page components (Table, Back and Submit button)
+- Success page
+
+### Update
+-  formik submit button
+
+
 09/14/2022
 ### Added
 - forms validation with stylings